edit/delete workout component bug fixes

+++ client/src/Components/DeleteWorkout.js
@@ -8,7 +8,8 @@ class DeleteWorkout extends Component {
 
         this.state = {
           workoutId:'',
-          trainingPlanId:''
+          trainingPlanId:'',
+          errors: []
         }
 
         this.onChange = this.onChange.bind(this);
@@ -22,9 +23,24 @@ class DeleteWorkout extends Component {
     }
 
     deleteWorkout() {
-      axios.delete('/workouts/' + this.state.workoutId)
-          .then(() => { window.location.reload(); })
-          .catch(err => console.log(err));
+      let errorsList = [];
+
+      if(this.state.trainingPlanId === '') {
+        errorsList.push('You must select a training plan');
+      }
+      if(this.state.workoutId === '') {
+        errorsList.push('You must select a workout to delete');
+      }
+
+      if(errorsList.length) {
+        this.setState({ errors: errorsList });
+      }
+
+      if(errorsList.length === 0) {
+        axios.delete('/workouts/' + this.state.workoutId)
+            .then(() => { window.location.reload(); })
+            .catch(err => console.log(err));
+      }
     }
 
      onChange(e) {
@@ -36,7 +52,21 @@ class DeleteWorkout extends Component {
 
       return(
         <div>
+          
+          <div className="modal-header">
+            <h5 className="modal-title">Delete Workout</h5>
+            <button type="button" className="close" data-dismiss="modal" aria-label="Close">
+              <span aria-hidden="true">&times;</span>
+            </button>
+          </div>
+          
           <div className="modal-body">
+            {this.state.errors.map( (error, index) => 
+              <div key={index} className="alert alert-danger" role="alert">
+                {error}
+              </div>
+            )}
+
             <label htmlFor="trainingPlan"><b>Training Plan to delete workout from:</b></label>
             <select id="inputState" 
               name="trainingPlanId" 
@@ -44,7 +74,7 @@ class DeleteWorkout extends Component {
               value={this.state.trainingPlanId ? this.state.trainingPlanId : ''} 
               onChange={this.onChange}
             >
-            <option defaultValue>Choose...</option>
+            <option value={''}>...</option>
              {this.props.trainingPlans.data.filter( (plan)=>plan.active ==true).map( (plan, index) =>
                 <option key={index} value={plan._id}>{plan.name}</option>
                 )
@@ -61,7 +91,7 @@ class DeleteWorkout extends Component {
                 value={this.state.workoutId ? this.state.workoutId : ''} 
                 onChange={this.onChange}
               >
-                <option selected>Choose...</option>
+                <option value={''}>...</option>
                 { this.props.workouts.data.filter( (workout) => workout.trainingPlan === this.state.trainingPlanId)
                   .map( (workout, index) => 
                   <option key={index} value={workout._id}>{workout.name}</option>
@@ -73,8 +103,8 @@ class DeleteWorkout extends Component {
           </div>
         
           <div className="modal-footer">
-            <button type="button" onClick={() => this.clearState() } className="btn btn-secondary btn-sm" data-dismiss="modal">Cancel</button>
-            <button type="button"  onClick={() => this.deleteWorkout() } className="btn btn-danger btn-sm">Delete Workout</button>
+            <button type="button" onClick={() => this.clearState() } className="btn btn-secondary" data-dismiss="modal">Cancel</button>
+            <button type="button"  onClick={() => this.deleteWorkout() } className="btn btn-danger">Delete Workout</button>
           </div>
 
         </div>
